Refactored auth and nav components
* Adjusted code to reflect new shared context
