+++ client/src/components/App.jsx
@@ -8,14 +8,15 @@ import {
   Text,
   Heading,
 } from '@chakra-ui/react';
-import LoginOption from './LoginOption';
+import LoginOption from './Auth/LoginOption';
 import ColorModeSwitcher from './ColorModeSwitcher';
 import PageControl from './PageControl';
 import { useAuthState } from 'react-firebase-hooks/auth';
-import { auth } from './Auth';
+import { auth } from './Auth/Auth';
 import ChatBar from './ChatBar';
 import NavButton from './NavButton';
 import { UseContextAll } from './ContextAll';
+import Welcome from './Welcome';
 
 import styles from '../style.css';
 
@@ -26,7 +27,7 @@ export default function App() {
     mainPage,
     setMainPage,
     userInfo,
-    setUserInfo
+    setUserInfo,
   } = UseContextAll();
 
   return (
@@ -46,15 +47,19 @@ export default function App() {
         {(() => {
           switch (mainPage) {
             case 'login':
-              return <LoginOption
-                        setMainDisplay={setMainPage}
-                        setUserID={setUserID}
-                        user={user}
-                        loading={loading}
-                        error={error}
-                      />;
+              return (
+                <LoginOption
+                  setMainPage={setMainPage}
+                  setUserID={setUserID}
+                  user={user}
+                  loading={loading}
+                  error={error}
+                />
+              );
             case 'pages':
               return <PageControl setMainDisplay={setMainPage} user={userInfo} />;
+            case 'welcome':
+              return <Welcome />;
             default:
               return <PageControl setMainDisplay={setMainPage} user={userInfo} />;
               // return <Logo/>; //  or som kind of load screen. This for option loading page
